refactoring: move ducks to separate directories
