+++ src/configureStore.js
@@ -1,12 +1,12 @@
-import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
-import { routerReducer, routerMiddleware } from 'react-router-redux';
-import { reducer as toastrReducer } from 'react-redux-toastr';
+import {createStore, combineReducers, applyMiddleware, compose} from 'redux';
+import {routerReducer, routerMiddleware} from 'react-router-redux';
+import {reducer as toastrReducer} from 'react-redux-toastr';
 import thunk from 'redux-thunk';
 
-import { branch as errorBranch, reducer as errorReducer } from './models/error';
-import { branch as usersBranch, reducer as usersReducer } from './models/users';
-import { branch as routerBranch } from './models/router';
-import { branch as toastrBranch } from './models/toastr';
+import {branch as errorBranch, reducer as errorReducer} from './models/error/error';
+import {branch as usersBranch, reducer as usersReducer} from './models/users/users';
+import {branch as routerBranch} from './models/router/router';
+import {branch as toastrBranch} from './models/toastr/toastr';
 
 
 let composeEnhancers = compose;
